feat: structured billing access log
